requirements.txt
gspread==6.1.4
google-auth==2.35.0
python-dateutil==2.9.0.post0